Create a config class
The class load a config file.
This config file can be in the default directory as config.ini.
The config file can although be loaded as a path.

If a scope is set the config var will be overwritten.
In the class are the default values.

Test with composer run tests.